Backend/requirements.txt
fastapi
uvicorn
motor
pydantic
email-validator
passlib[bcrypt]
bcrypt==4.1.2
python-jose[cryptography]
python-multipart
razorpay
python-dotenv
pyswisseph
reportlab
